DRAFT - Debugging ReSTIR GI bias when reusing specular sample

+++ src/Device/kernels/ReSTIR/GI/SpatialReuse.h
@@ -92,7 +92,7 @@ GLOBAL_KERNEL_SIGNATURE(void) inline ReSTIR_GI_SpatialReuse(HIPRTRenderData rend
 
 	ReSTIRSpatialResamplingMISWeight<ReSTIR_GI_BiasCorrectionWeights, /* IsReSTIRGI */ true> mis_weight_function;
 
-	float DEBUG_SELECTGED_JACBOAIN = 0.0f;
+	int DEBUG_SELECTGED_JACBOAIN = 0;
 
 	int reused_neighbors_count = render_data.render_settings.restir_gi_settings.common_spatial_pass.reuse_neighbor_count;
 	int start_index = 0;
